Tipos de retorno dos gets definidos como int. Mesmo uma PK sendo short, para querys é mais seguro se utilizar o tipo int.

+++ trunk/biblioteca-eeep/src/fvsosp/biblioteca/Biblioteca.java
@@ -55,7 +55,7 @@ public class Biblioteca implements Serializable{
     private List<Acervo> acervos;
 
     /*** @retorna o ID da Biblioteca ***/
-    public short getIdBiblioteca() {
+    public int getIdBiblioteca() {
         return idBiblioteca;
     }
